Resolves #164
